[WarpSpec] add support for multiple channels sharing the same smem (#9)
location

Summary: We already have channelsGroupedByProducers and
channelsGroupedByConsumers. For one-producer-multi-consumer mode, a
single buffer will be used, channelsGroupedByProducers is used for this.
channelsGroupedByConsumers is to minimize the insertion of sync
primitives, a single set of communication ops will be inserted.

For this patch, we want to share the same smem location for multiple
channels that are live in different loop nests. We add
allocation.shareGroup attributes to the local_allocs corresponding to
channels that reuse the same smem location.

In order to reuse the same smem location, we update bufferIdx and phase
through all the loop nests that share smem locations. We handle the
following cases:
```
for # persistent loop
  for # can be nested under if
  for # can be nested under if
Or
for # can be nested under if
for # can be nested under if
Or
for # persistent loop
  for # can be nested under if
```

The generated code will look like
```
for(accumLoopCount)
  t1 = IfOp
    forOp # loop A
    tmpIdx = accumLoopCount + numStepsA
    yield tmpIdx
    else yield accumLoopCount
  t2 = IfOp
    forOp # loop B
    tmpIdx = t1 + numStepsB
    yield tmpIdx
    else yield t1
  yield t2 for accumLoopCount
```