Improve data partitioning to handle tmem_load/alloc (#42)
tmem_load, tmem_alloc can have side effects so we can no longer depend
on populateForOpDeadArgumentElimination/ForOp cleanup to remove the
original ops that should be replaced with the sliced versions. We
removed the original ops by walking through the function op and erasing
ops in partitionScheme.ops.

Also this patch tries to correctly handle the slicing of
tmem_load/alloc: an extra convert_layout is added to convert the blocked
layout that is used by tmem_load/alloc to be compatible with the sliced
tmem.

There are a few places where we need to keep attributes. These changes
exist in the ws branch but in a different file.