Attaching load type to tokens for accurate load type inference (#43)
Summary:
Inferring the type of load a token tracks in wslowering is expansive and
often leads to wrong results, since there is no explicitly connection
between the two. I'm fixing this by attaching the load op type to the
token when it is created. This fixes a deadlock issue where wrong load
type was inferred and thus wrong number of threads were expected by the
token.

Test Plan:

Reviewers:

Subscribers:

Tasks:

Tags: