Add a fix pass to rewrite the incorrect bar.sync 0 in WS (#61)
This PR adds a pass to rewrite the incorrect bar.sync 0 in WS. It
performs a forwardSlice analysis on getting users of the
`CanonicalWarpIdOp`, which is the by-product of the warp specialization
support. We use that OP as an anchor. After we get all users from
`CanonicalWarpIdOp`, we perform a dominance analysis to collect all
WSLocal basic blocks. In each WSLocal code block, we rewrite any
`bar.sync 0` to `bar.sync BARID, 128` as a goal to reuse the barrier id.
We have a hard-coded warp group size as 4 warps (128 threads). Here
BARID is the barrier id, which we try to reuse as much as possible
(BARID must < 16). A similar approach based on `PartitionedRegion` has
been implemented in OAI so after the WS code merge we could drop this
fix pass. All intended `gpu::barrierOp` (`bar.sync 0`) in the triton
compiler pipeline in WS passes should add the attributes: `barid` and
`threadcount` to enjoy the direct PTX inline code generation in the LLVM
conversion, rather than relying on the `nvvm` conversion.

---------

Co-authored-by: Yuanwei Fang <fywkevin@fb.com>