[WS] Recursive search for used args during code split (#68)
During code split we need to know if a block arg is supposed to be
duplicated into a specific task region. It isn't straightforward to get
the task id of the arg from on IR as such id is not physically stored
anywhere. All we have is the task id of its parent op which is likely a
control op and shared by all tasks. The existing solution is to compute
the task id for the arg based on its initial values. However, a block
arg of an inner loop can be initialized by a block arg of an outer loop,
which needs recursive upward tracking. We are doing it in another way
that is to recursively downward track the user of the arg.