fix codepartition issue
Summary: Do not create a channel if output of a dot is used as initArg
of a ForOp and the actual argument is only used by the same taskId.
Also count IfOp as control depth so we can correctly handle
If
  For
  For
Two ForOps will have the same control depth.

Test Plan:

Reviewers:

Subscribers:

Tasks:

Tags: