fix channels of operand A
Summary:
We can create two channels for a gen5, one is for the accumulator, the
other is for operand A when it is in tmem.

When operand A is in smem, it will probably be rooted from local_alloc
when taskId of local_alloc is different from the user's taskId.

Test Plan:

Reviewers:

Subscribers:

Tasks:

Tags: