Improve how dynamic register reallocation is implemented (#6694)
This PR generally improves how the compiler handles dynamic register
reallocation for warp specialization.

* Codegen generates `setmaxnreg` both at the entry and exit of the
partition regions, even when the number of registers does not change.
This somehow makes `ptxas` behave much better, allowing the registers
allocated to the load and MMA partitions to drop to `24` as they should
be. This should improve register pressure across the board for warp
specialized kernels.
* The maximum number of warpgroups is computed across the whole program
and each `ttg.warp_specialize` is padded to it. This ensures all warps
are always present to surrender registers. This primarily improves the
layering between partitioning in the middle end (no longer need the
"extra empty warp" hack).
* Handle TMEM ops when relayout'ing the IR (they require a minimum of 4
warps). Thankfully, the TMEM compatible distributed layout can always be
inferred for these ops.