[PROTON-DEV] Chrome Trace Writer (#6772)
This PR adds the trace writer utilities for the chrome tracing.
Specifically, it supports dumping a json trace in chrome trace format
for a single stream with multiple kernels. Each kernels has a constant
gap `kKernelTimeGap`. Due to the limitation of chrome trace, each pid
line is a CTA triplet (kernel name, Processor ID, CTA ID). Inside PID,
each TID is a portion of warp events recorded. Here we say portion
because we apply a line splitting to avoid chrome tracing corrupts
overlapping events. Any overlapping events in the same warp would be
placed in a separate TID line. In this implementation, each CTA applies
an independent start event offset. As a result, it is only safe to
compare events within the same CTA.