Better warp specialization register reallocation (#6877)
FIXME: ignore change to PartitionScheduling.cpp
This is the last set of changes split from #6760

This moves where the registers are reallocated into better spots. It
also causes the worker partitions to immediately give up a bunch of
registers when they aren't active to the default warp group, so that it
has more registers to execute the "synchronous" parts of the code. This
is useful when there are many worker warps and the default warp group
does not get many registers at the start of the kernel (maxnreg is low).