[BACKEND] Add a new pass to insert fence.proxy.async for write after read hazard (#7262)
When loading data from smem using the generic proxy then writing using
the async proxy we need to insert a fence before writing. This case was
being missed in our current analysis. This commit hads a new pass to
conservatively insert fences in this kind of scenario. We still need the
existing insertFence pass as it can be smarter about fence insertion
because it runs before lowering of the structured control flow. This
pass is meant to be more conservative while the previous pass can
optimize better.
Didn't see significant perf regressions so far.