Fix compiler warnings from Membar virtual function change (#7324)
Fixes the `'mlir::MembarOrFenceAnalysis' has virtual functions but
non-virtual destructor` warning that occurs due to adding a virtual
function to the `MembarOrFenceAnalysis` base class. This error fails our
`-Werror` check because we do not disable `non-virtual-dtor`.

I also fixed `warning: field 'numCols' will be initialized after field
'numRows'` since it showed up in my error trace, but this is not a
warning that prompts errors, so there may be other locations where this
warning is occurring.