[TLX] Put an attr in module op to indicate TLX is active (#198)
To reliably disable certain optimization passes, we detect TLX ops (or
certain TTG/TTNG ops introduced to TTIR by TLX) and indicate hasTLX as a
module op attribute. This is being done in Fixup pass. Other passes will
be able to just read the attr later to decide whether they want to
disable some parts of itself.

The list of Ops that render the kernel as "having TLX" is enumerated
from builder functions in `triton_tlx.cc` and anything that's in `ir.cc`
is not included in the list.

Testing:
- `make test-lit`
- `pytest -vs python/test/unit/language/test_tlx.py`, only one
pre-existing failure (test_async_tasks)
- `python python/tutorials/03-matrix-multiplication.py`. The kernel runs
fine and I verified there's no such attr in ttir/ttgir.