[TLX] Fix issues where default num_warps cannot be one with async_task (#296)
When default (program-level) `num_warps` is set to one, the compilation
would fail with a tmem load or store used in a task region. This is
because tmem load/store requires a region of 4 or 8 warps, and during
the dry run of async_task compilation (mainly for computing variable
captures), the global `num_warps` instead of region `num_warps` is used.
I'm fixing it by doing two dry-runs. The first one is to compute region
information, followed by WS op creation, and the variable capturing is
done in the second dry-run where region `num_warps` is available.