[WS] Moving buffer creation earlier (#335)
Moving buffer creation logics earlier so that smem/tmem allcoation will
be hoisted to the top of a kernel, for a better handling for downstream
passes.

A smem channel like below 

```
    scf.for %arg25 = %c0_i32 to %9 step %c1_i32  : i32 {
      %30 = tt.descriptor_load %arg4[%29, %c0_i32] {ttg.partition = 4 : i32} 
      %31 = ttg.local_alloc %30 {ttg.partition = 4 : i32} 
       %59 = ttng.tc_gen5_mma %31, %56, %result[%arg32], %false, %true {loop.cluster = 0 : i32, loop.stage = 2 : i32, tt.self_latency = 1 : i32, ttg.partition = 3 : i32}
```

now becomes:

```
  %0 = "ttg.local_alloc"() : () 
  scf.for %arg25 = %c0_i32 to %9 step %c1_i32  : i32 {
     %51 = "tt.descriptor_load"(%arg4, %50, %16) <{cache = 1 : i32, evict = 1 : i32}> {ttg.partition = 4 : i32} : 
     %52 = "arith.constant"() <{value = 0 : i32}> 
     %53 = "ttg.memdesc_index"(%0, %52) 
     "ttg.local_store"(%51, %53) 
     %91 = "ttng.tc_gen5_mma"(%53, %87, %58#0, %arg32, %11, %12)

```

TMEM channel is handled in a similar way.

Register channels will be handled in another PR, where we need to to
decide on whether TMEM or SMEM buffers are needed to back them up.