Handle producers in different blocks for consumer channel grouping (#350)
We currently merge channels with the same consumer op into the same
channel group, even if they fail the channelCanBeMerged check.

This means that in theory, if we had two channels with producers in
different blocks, they could end up in the same channel group if they
have the same consumer operation. This would break in the next section
of the code, where we attempt to sort the channels within each group by
the program order of their producer ops.

To address this, always create a new channel group when we fail to merge
a channel into any existing group.